Change <popup> to LayoutObjectIsNeeded from AdjustStyleForHTMLElement
This approach gives the same end-result: popups aren't shown if
their 'open' property is false. But per this comment [1], removing
the entire layout object is a better way to accomplish this.

Note that Element::SetIsInTopLayer() already calls
SetForceReattachLayoutTree() at [2], so there's no need to call
it here in MarkStyleDirty(). The call to SetNeedsStyleRecals()
does seem to be necessary, however, or no popup is rendered.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2648613/8/third_party/blink/renderer/core/html/resources/html.css#b1237
[2] https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/core/dom/element.cc;l=5366;drc=699e1bd813aa71e101fc20a6ded0d23a277181e4

Bug: 1170977
Change-Id: I3368af36631816d8377ab3ba0ea052f2ac1e138a
Cq-Do-Not-Cancel-Tryjobs: true